pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]
